requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.0
django-filter>=23.0
python-dotenv>=1.0.0
Pillow>=10.0.0  # for ImageField
whitenoise>=6.5.0  # for serving static files 